+++ packages/gatsby-config/CHANGELOG.md
@@ -3,12 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.2.14](https://github.com/kamontat/kcutils/compare/@kcutils/gatsby-config@0.2.13...@kcutils/gatsby-config@0.2.14) (2020-09-11)
+## [0.2.16-beta.0](https://github.com/kamontat/kcutils/compare/@kcutils/gatsby-config@0.2.14...@kcutils/gatsby-config@0.2.16-beta.0) (2020-09-15)
+
+
+### Bug Fixes
+
+* **core:** compile error due to @kcutils/logger changes ([a107abe](https://github.com/kamontat/kcutils/commit/a107abed82f8740ae824b38994dfc4dde9ba66ad))
+
+
+
+
+
+## [0.2.14-beta.4](https://github.com/kamontat/kcutils/compare/@kcutils/gatsby-config@0.2.14-beta.3...@kcutils/gatsby-config@0.2.14-beta.4) (2020-09-10)
+
+**Note:** Version bump only for package @kcutils/gatsby-config
+
+
+## [0.2.14-beta.3](https://github.com/kamontat/kcutils/compare/@kcutils/gatsby-config@0.2.14-beta.2...@kcutils/gatsby-config@0.2.14-beta.3) (2020-09-10)
+
+**Note:** Version bump only for package @kcutils/gatsby-config
+
+
+## [0.2.14-beta.2](https://github.com/kamontat/kcutils/compare/@kcutils/gatsby-config@0.2.14-beta.1...@kcutils/gatsby-config@0.2.14-beta.2) (2020-09-09)
 
 **Note:** Version bump only for package @kcutils/gatsby-config
 
 
 
+
+
+## [0.2.14-beta.1](https://github.com/kamontat/kcutils/compare/@kcutils/gatsby-config@0.2.14-beta.0...@kcutils/gatsby-config@0.2.14-beta.1) (2020-09-09)
+
+**Note:** Version bump only for package @kcutils/gatsby-config
+
+
+
+
+
+## [0.2.14-beta.0](https://github.com/kamontat/kcutils/compare/@kcutils/gatsby-config@0.2.13...@kcutils/gatsby-config@0.2.14-beta.0) (2020-09-09)
+
+
+### Bug Fixes
+
+* **core:** compile error due to @kcutils/logger changes ([a107abe](https://github.com/kamontat/kcutils/commit/a107abed82f8740ae824b38994dfc4dde9ba66ad))
 
 
 ## [0.2.13](https://github.com/kamontat/kcutils/compare/@kcutils/gatsby-config@0.2.12...@kcutils/gatsby-config@0.2.13) (2020-09-05)